Update UI QMM012

+++ nts.uk/ctx/pr/core/nts.uk.ctx.pr.core.app/src/main/java/nts/uk/ctx/pr/core/app/command/itemmaster/DeleteItemMasterCommandHandler.java
@@ -7,6 +7,8 @@ import javax.inject.Inject;
 import javax.transaction.Transactional;
 
 import lombok.val;
+import nts.arc.error.BusinessException;
+import nts.arc.error.RawErrorMessage;
 import nts.arc.layer.app.command.CommandHandler;
 import nts.arc.layer.app.command.CommandHandlerContext;
 import nts.uk.ctx.pr.core.dom.itemmaster.ItemMasterRepository;
@@ -52,6 +54,8 @@ public class DeleteItemMasterCommandHandler extends CommandHandler<DeleteItemMas
 		val companyCode = AppContexts.user().companyCode();
 		val categoryAtr = context.getCommand().getCategoryAtr();
 		val itemCode = context.getCommand().getItemCode();
+		if (!this.itemMasterRepository.find(companyCode, categoryAtr, itemCode).isPresent())
+			throw new BusinessException(new RawErrorMessage("更新対象のデータが存在しません。"));
 		this.itemMasterRepository.remove(companyCode, categoryAtr, itemCode);
 		// remove all sub item linked with item master through the item code
 		switch (categoryAtr) {
@@ -72,30 +76,41 @@ public class DeleteItemMasterCommandHandler extends CommandHandler<DeleteItemMas
 
 	private void deleteItemSalary(String companyCode, CommandHandlerContext<DeleteItemMasterCommand> context) {
 		val itemCode = context.getCommand().getItemCode();
+		if (!this.itemSalaryRespository.find(companyCode, itemCode).isPresent())
+			throw new BusinessException(new RawErrorMessage("更新対象のデータが存在しません。"));
 		this.itemSalaryRespository.delete(companyCode, itemCode);
+		if (this.itemSalaryPeriodRepository.find(companyCode, itemCode).isPresent())
 		this.itemSalaryPeriodRepository.delete(companyCode, itemCode);
 		// Then remove itemBD
 		List<ItemSalaryBD> itemBDList = this.itemSalaryBDRepository.findAll(companyCode, itemCode);
 		for (ItemSalaryBD itemBD : itemBDList) {
+			if (!this.itemSalaryBDRepository.find(companyCode, itemCode, itemBD.getItemBreakdownCode().v()).isPresent())
+				throw new BusinessException(new RawErrorMessage("更新対象のデータが存在しません。"));
 			this.itemSalaryBDRepository.delete(companyCode, itemBD.getItemCode().v(),
 					itemBD.getItemBreakdownCode().v());
 		}
-
 	}
 
 	private void deleteItemDeduct(String companyCode, CommandHandlerContext<DeleteItemMasterCommand> context) {
 		val itemCode = context.getCommand().getItemCode();
+		if (!this.itemDeductRespository.find(companyCode, itemCode).isPresent())
+			throw new BusinessException(new RawErrorMessage("更新対象のデータが存在しません。"));
 		this.itemDeductRespository.delete(companyCode, itemCode);
+		if (this.itemDeductPeriodRepository.find(companyCode, itemCode).isPresent())
 		this.itemDeductPeriodRepository.delete(companyCode, itemCode);
 		List<ItemDeductBD> itemBDList = this.itemDeductBDRepository.findAll(companyCode, itemCode);
 		for (ItemDeductBD itemBD : itemBDList) {
-			this.itemDeductBDRepository.delete(companyCode,itemBD.getItemCode().v(), itemBD.getItemBreakdownCode().v());
+			if (!this.itemDeductBDRepository.find(companyCode, itemCode, itemBD.getItemBreakdownCode().v()).isPresent())
+				throw new BusinessException(new RawErrorMessage("更新対象のデータが存在しません。"));
+			this.itemDeductBDRepository.delete(companyCode, itemBD.getItemCode().v(),
+					itemBD.getItemBreakdownCode().v());
 		}
-
 	}
 
 	private void deleteItemAttend(String companyCode, CommandHandlerContext<DeleteItemMasterCommand> context) {
 		val itemCode = context.getCommand().getItemCode();
+		if (!this.itemAttendRespository.find(companyCode, itemCode).isPresent())
+			throw new BusinessException(new RawErrorMessage("更新対象のデータが存在しません。"));
 		this.itemAttendRespository.delete(companyCode, itemCode);
 
 	}
